store: add workerEndpoints for composerApi
